feat: useFetch hook
Created use fetch hook with get, post, put and delete methods to make request easier.

"feat #6"
