+++ src/hooks/index.ts
@@ -1,1 +1,1 @@
-export {default as useApi} from "./useApi";
+export {default as useFetch} from "./useFetch";
