--- src/interfaces/api-config.interface.ts
@@ -1,8 +1,0 @@
-export interface IApiConfig {
-    url: string;
-    method: string;
-    data?: any;
-    headers?: { [key: string]: string};
-}
-
-export default IApiConfig;
