--- src/interfaces/api-params.interface.ts
@@ -1,6 +1,0 @@
-export interface IApiParams {
-    baseUrl: string;
-    bearerToken?: string // Optional bearer token parameter
-}
-
-export default IApiParams;
