+++ src/interfaces/index.ts
@@ -1,4 +1,3 @@
-export {default as IApiConfig} from "./api-config.interface";
-export {default as IApiParams} from "./api-params.interface";
+export {default as IApiRequest} from "./api-request.interface";
 export {default as IApiResponse} from "./api-response.interface";
 export {default as IApiStatus} from "./api-status.interface";
